chore(createCommentForm): use validation, and create comment hooks,and rename the dir
